requirements.txt
docker_py==1.10.6
requests==2.28.1
run==0.2
setup==1.0
urllib3==1.26.13
